+++ extension/src/goToolsInformation.ts
@@ -85,6 +85,15 @@ export const allToolsInformation: { [key: string]: Tool } = {
 		description: 'Linter',
 		minimumGoVersion: semver.coerce('1.20')
 	},
+	'golangci-lint-v2': {
+		name: 'golangci-lint-v2',
+		importPath: 'github.com/golangci/golangci-lint/v2/cmd/golangci-lint',
+		modulePath: 'github.com/golangci/golangci-lint/v2',
+		replacedByGopls: false,
+		isImportant: true,
+		description: 'Linter',
+		minimumGoVersion: semver.coerce('1.23')
+	},
 	'revive': {
 		name: 'revive',
 		importPath: 'github.com/mgechev/revive',